renamed parser library file
